requirements.txt
# Core Flask
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.8.0

# ORM / DB
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

# Forms / validation
WTForms==3.1.2
email-validator==2.2.0

# Document processing (local fallback)
PyPDF2==3.0.1
pdfminer.six==20231228
python-docx==1.1.2
Pillow==10.4.0

# Security
cryptography==43.0.3
itsdangerous==2.2.0
bleach==6.2.0
python-magic==0.4.27

# Utilities
python-dotenv==1.0.1
Werkzeug==3.0.6

# AWS SDK — S3, SES, SNS, Textract, Comprehend, CloudWatch, Cognito
boto3==1.35.54
botocore==1.35.54

# Task queue (Celery with Redis or SQS broker)
celery==5.4.0
redis==5.2.0

# WSGI server
gunicorn==23.0.0

# PDF report export (for evaluation reports)
WeasyPrint==62.3
